feat: add button for comment input
